requirements.txt.txt
flask
flask-cors
pytesseract
pillow
opencv-python
pdf2image
python-docx
spacy